first_controller_draft

+++ db/migrate/20231116095807_create_movies.rb
@@ -4,7 +4,7 @@ class CreateMovies < ActiveRecord::Migration[7.1]
       t.string :title
       t.text :overview
       t.string :poster_url
-      t.integer :rating
+      t.float :rating
 
       t.timestamps
     end
